backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.6
packaging==24.2
PyJWT==2.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1